fix(sweep-task): fix import errors
