+++ develop_tools/test/test_explore_hard_task.py
@@ -1,10 +1,11 @@
 import unittest
+
+import module.ExploreTasks.TaskUtils
 from main import Main
-from module.ExploreTasks.TaskUtils import to_mission_info, to_region, execute_grid_task, employ_units, \
-    convert_team_config
-from core import image, picture
+from module.ExploreTasks.TaskUtils import to_mission_info, to_region, execute_grid_task
 from module.ExploreTasks.explore_task import validate_and_add_task
-from module import normal_task, hard_task, main_story
+from module import main_story
+from module.ExploreTasks import normal_task, sweep_task
 from core.Baas_thread import Baas_thread
 from core.config.config_set import ConfigSet
 
@@ -53,7 +54,7 @@ class TestExploreNormalTask(unittest.TestCase):
                 self.logger.info(f"--- Start exploring H{taskName}({taskDataName}) ---")
 
                 mission_los = [249, 363, 476]
-                hard_task.to_hard_event(self.baas_thread, True)
+                module.ExploreTasks.TaskUtils.to_hard_event(self.baas_thread, True)
                 if not (to_region(self.baas_thread, region, False) and
                         to_mission_info(self.baas_thread, mission_los[mission - 1])):
                     self.logger.error(f"Skipping task {taskName} since it's not available.")
@@ -66,6 +67,6 @@ class TestExploreNormalTask(unittest.TestCase):
                 main_story.auto_fight(self.baas_thread)
 
                 # skip unlocking animation by switching
-                normal_task.to_normal_event(self.baas_thread, True)
-                hard_task.to_hard_event(self.baas_thread, True)
+                module.ExploreTasks.TaskUtils.to_normal_event(self.baas_thread, True)
+                module.ExploreTasks.TaskUtils.to_hard_event(self.baas_thread, True)
         return True
