+++ module/ExploreTasks/__init__.py
@@ -1,2 +1,1 @@
-from module.ExploreTasks import TaskUtils
-from module.ExploreTasks import explore_task
+from module.ExploreTasks import TaskUtils, explore_task, sweep_task
