+++ module/ExploreTasks/explore_task.py
@@ -1,7 +1,8 @@
+import module.ExploreTasks.TaskUtils
 from core import color, image, picture
-from module import main_story, normal_task, hard_task
+from module import main_story
 from module.ExploreTasks.TaskUtils import to_mission_info, execute_grid_task, get_challenge_state, \
-    employ_units, get_stage_data, convert_team_config, to_region
+    employ_units, get_stage_data, convert_team_config, to_region, to_hard_event, to_normal_event
 
 
 def validate_and_add_task(self, task: str, tasklist: list[tuple[int, int, dict]],
@@ -132,8 +133,7 @@ def explore_normal_task(self):
     teamConfig = convert_team_config(self)
 
     for task in tasklist:
-        region = task[0]
-        mission = task[1]
+        region, mission = task[0], task[1]
 
         # skip navigate to mission if previous task is already finished
         skip_navigate = False
@@ -143,7 +143,7 @@ def explore_normal_task(self):
             self.logger.info(f"--- Start exploring {taskName}({taskDataName}) ---")
 
             if not skip_navigate:
-                normal_task.to_normal_event(self, True)
+                to_normal_event(self, True)
                 if not to_region(self, region, True):
                     self.logger.error(f"Skipping task {taskName} since it's not available.")
                     continue
@@ -201,8 +201,8 @@ def explore_normal_task(self):
                 if self.config.manual_boss:
                     self.click(1235, 41)
             # skip unlocking animation by switching
-            hard_task.to_hard_event(self, True)
-            normal_task.to_normal_event(self, True)
+            to_hard_event(self, True)
+            to_normal_event(self, True)
     return True
 
 
@@ -232,8 +232,7 @@ def explore_hard_task(self):
     self.logger.info("}")
 
     for task in tasklist:
-        region = task[0]
-        mission = task[1]
+        region, mission = task[0], task[1]
 
         # skip navigate to mission if previous task is already finished
         skip_navigate = False
@@ -243,7 +242,7 @@ def explore_hard_task(self):
 
             mission_los = [249, 363, 476]
             if not skip_navigate:
-                hard_task.to_hard_event(self, True)
+                to_hard_event(self, True)
                 if not (to_region(self, region, False) and to_mission_info(self, mission_los[mission - 1])):
                     self.logger.error(f"Skipping task {taskName} since it's not available.")
                     continue
@@ -262,6 +261,6 @@ def explore_hard_task(self):
             if self.config.manual_boss:
                 self.click(1235, 41)
             # skip unlocking animation by switching
-            normal_task.to_normal_event(self, True)
-            hard_task.to_hard_event(self, True)
+            to_normal_event(self, True)
+            to_hard_event(self, True)
     return True
